Fix duplicate headers causing invalid requests (#10)
Fix for: Azure/azure-iot-pcs-remote-monitoring-dotnet#12

In some situations the proxy is sending the Content-Type header with a wrong value, e.g.
`Content-Type: application/json, application/json` due to the way that ASP.NET HTTP client manages headers partially on the request and partially on the content. The fix tracks the headers added to the content to avoid the duplication.